split mat and ext
